backend/requirements.txt
fastapi
uvicorn[standard]

pyjwt
python-dotenv
requests

google-cloud-firestore